requirements.txt
torch==2.2.0
torchvision==0.17.0
transformers==4.36.2
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
tqdm==4.66.1
numpy==1.26.4
Pillow==10.2.0
scipy==1.12.0
joblib==1.3.2